feat: useSearch hook
